Title: 낮은 지점들, Time: 116ms, Memory: 2MB, Status: Passed - CodeTree

+++ 231222/낮은 지점들/lowest-points.cpp
@@ -18,7 +18,7 @@ int main() {
             m[x] = y;
         }
     }
-    int answer = 0;
+    long long answer = 0;
     for (auto i : m) {
         answer += i.second;
     }
